refactor(synapsys): extract subagent-match + emit-matched helpers to satisfy max-lines (GH-497)
After merging GH-497 onto current main, plugins/synapsys/hooks/synapsys.js
reached 484 lines, exceeding the 400-line max-lines static gate. Extract the
GH-497 subagent prompt-scope propagation helpers (collectSubagentMatches,
buildSubagentSelectOpts, unionByName, computeMatched) into hooks/lib/subagent-matches.js
and the telemetry/expectation recorder (emitMatched) into hooks/lib/emit-matched.js.
synapsys.js is now 384 lines. No behavior change; dispatcher integration tests
(dispatcher-pretool-injection) pass 10/10.
